+++ CHANGELOG.md
@@ -37,6 +37,7 @@
   - Add `ItemFact::set_context_ref` and `ItemFact::set_unit_ref`
   - Add `TypedInstanceDocument` and `XbrlError::ValueError`
   - Make `resolve_instance` public
+  - Parse `parts` from reference linkbase
 - changed
   - Parse `ComplexType`s recursively
   - Rename `InstanceParser::parse_instance` to `InstanceParser::parse`